refactor: Major frontend restructure and CalendarComponent refactor
- Massively refactor CalendarComponent for better maintainability
- Reorganize component structure with improved folder hierarchy:
  - Move dialogs to dedicated Dialogs/ folder
  - Reorganize forms components into Forms/ subfolders (CourseList, CourseSearch, Settings)
  - Create dedicated Calendar/ folder with hooks and utilities
  - Move UI components to UI/ folder
  - Move export components to Export/ folder
- Add custom hooks for calendar functionality:
  - useCalendarEventHandlers, useTouchEvents, useEventBusHandlers
  - useTimetableManagement, useCalendarEvents
- Split calendar utilities into focused modules:
  - calendarConfigUtils, calendarUtils, calendarViewUtils
  - courseTimelineUtils, eventHandlerUtils
- Fix JSX imports in utility files to resolve build errors
- Fix ICS export undefined values in BYDAY fields
- Remove console.logs and apply consistent code formatting
- Remove outdated file structure documentation
- Improve code organization and separation of concerns
