--- docs/FileStructure.md
@@ -1,61 +1,0 @@
-# File Structure Documentation
-
-This document explains the purpose of each file and directory in the repository.
-
-## Root Directory
-
-- [**docs/**:](#docs) This directory contains markdown files with different documentation to help you understand the layout and functionality of this application.
-
-- [**public/**:](#public) This directory typically contains static assets that are publicly accessible. In this project, it may include things like images, videos.
-
-- [**src/**:](#src) The main source code of the application resides here. It contains all React components, styles, and other code that powers the application.
-
-- **.eslintrc.cjs**: Configuration file for ESLint, a tool used to analyze code for potential errors and enforce coding standards. This file defines the rules and settings for ESLint within this project.
-
-- **.gitignore**: A file specifying which files and directories should be ignored by Git. This usually includes node_modules, build artifacts, and other temporary files that shouldn't be tracked in version control.
-
-- **README.md**: Contains an overview of the project, instructions on how to set up and run it, and any other relevant information for developers and users. It’s the first point of reference for anyone using or contributing to the project.
-
-- **index.html**:This file is the main HTML template for the application. Vite uses this file to serve the app during development and build processes.
-
-- **package-lock.json**: Automatically generated by npm, this file locks the version of every installed package, ensuring that the same versions are installed if someone else installs the project’s dependencies.
-
-- **package.json**: This file contains metadata about the project, including the list of dependencies, scripts to run tasks (like start, build, test), and other configurations. It is crucial for managing the project’s dependencies and scripts.
-
-- **vite.config.js**: The configuration file for Vite, a build tool that is optimized for frontend development. This file contains settings and plugins that Vite uses to bundle and serve the application.
-
-## /docs/
-
-The `docs/` directory contains markdown files with different documentation to help you understand the layout and functionality of this application.
-
-- **API.md**: The backend API endpoint documentation.
-
-- **FileStructure.md**: The file structure documentation.
-
-## /public/
-
-The `public/` directory typically contains static assets that are publicly accessible. In this project, it may include things like images, videos.
-
-- **BrockTimeTableTutorial.mp4**: The generator user tutorial video.
-
-- **favicon.ico**: Main react component wrapper for the entire application.
-
-## /src/
-
-The `src/` directory contains all the source code for the project.
-
-- **GeneratorPage/**: This directory contains all code, css, contexts and react components for the timetable generator page.
-
-- **SiteWide/**: This directory contains all code, css, contexts and react components for anything that may be needed across multiple pages of the application.
-
-- **GeneratorPage.jsx**: Main page react component wrapper for the generator page.
-
-- **Main.jsx**: Main react component wrapper for the entire application.
-
-### General Structure of Subfolders in /src/ (GeneratorPage/ & SiteWide/)
-
-- **components/**: This directory is dedicated to storing React components. Only `.jsx` files should be placed here. If a component is composed of sub-components, the sub-components will be organized within a folder named `"NameComponents/"` for better structure and clarity.
-
-- **scripts/**: This directory is dedicated to storing raw JavaScript code. Only `.js` files should be placed here.
-
-- **css/**: This directory is dedicated to storing css files. Only `.css` files should be placed here.
