+++ src/GeneratorPage/components/Dialogs/BlockedSlotsDialog.jsx
