+++ src/GeneratorPage/components/Dialogs/ChangelogDialogComponent.jsx
