+++ src/GeneratorPage/components/Dialogs/TruncationDialog.jsx
