+++ src/GeneratorPage/components/Forms/CourseList/CourseList.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import CourseListComponent from "../CourseListComponent";
-import BorderBox from "./BorderBox";
+import CourseListComponent from "./CourseListComponent";
+import BorderBox from "../../UI/BorderBox";
 
 export default function CourseList({
   addedCourses,
