+++ src/GeneratorPage/components/Forms/CourseSearch/AddButtonComponent.jsx
