+++ src/GeneratorPage/components/Forms/Settings/SortOptions.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import SortDropdown from "../SortDropdown";
-import BorderBox from "./BorderBox";
+import SortDropdown from "./SortDropdown";
+import BorderBox from "../../UI/BorderBox";
 
 export default function SortOptions({ sortChoice, handleSortChange }) {
   return (
