+++ src/GeneratorPage/components/UI/BorderBox.jsx
