+++ src/GeneratorPage/components/UI/SourceCode.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Box from "@mui/material/Box";
-import GitHubButton from "../GitHubButton";
+import GitHubButton from "./GitHubButton";
 import BorderBox from "./BorderBox";
 
 export default function SourceCode() {
