+++ src/GeneratorPage/components/index.jsx
@@ -1,9 +1,9 @@
 export { default as NavbarComponent } from "../../SiteWide/components/NavbarComponent";
 
-export { default as CalendarComponent } from "./CalendarComponent";
+export { default as CalendarComponent } from "./Calendar/CalendarComponent";
 
-export { default as CourseListComponent } from "./InputFormComponents/CourseListComponent";
+export { default as CourseListComponent } from "./Forms/CourseList/CourseListComponent";
 
-export { default as InputFormTopComponent } from "./InputFormTopComponent";
+export { default as InputFormTopComponent } from "./Forms/InputFormTopComponent";
 
-export { default as InputFormBottomComponent } from "./InputFormBottomComponent";
+export { default as InputFormBottomComponent } from "./Forms/InputFormBottomComponent";
